Move core modules into package
