+++ cli.py
@@ -3,8 +3,8 @@ import json
 import sys
 from pathlib import Path
 
-from src.ai_karen_engine import LLMOrchestrator
-from src.core.plugin_router import PluginRouter
+from ai_karen_engine import LLMOrchestrator
+from ai_karen_engine.core.plugin_router import PluginRouter
 
 
 CONFIG_PATH = Path("config") / "settings.json"
