+++ src/ai_karen_engine/__init__.py
@@ -3,8 +3,8 @@
 from .clients.slm_pool import SLMPool
 from .llm_orchestrator import LLMOrchestrator
 from .echocore.fine_tuner import NightlyFineTuner
-from src.core.model_manager import ModelManager
-from src.core.echo_core import EchoCore
+from ai_karen_engine.core.model_manager import ModelManager
+from ai_karen_engine.core.echo_core import EchoCore
 from src.clients.transformers.lnm_client import LNMClient
 from src.clients.nlp.basic_classifier import BasicClassifier
 from src.clients.nlp.spacy_client import SpaCyClient
