+++ src/ai_karen_engine/core/__init__.py
@@ -1,1 +1,6 @@
 """Core utilities for AI Karen engine."""
+
+from .mesh_planner import MeshPlanner
+from .gpu_training import gpu_optimized_train
+
+__all__ = ["MeshPlanner", "gpu_optimized_train"]
