+++ src/ai_karen_engine/core/automation_manager.py
