+++ src/ai_karen_engine/core/echo_core.py
