+++ src/ai_karen_engine/core/embedding_manager.py
