+++ src/ai_karen_engine/core/mesh_planner.py
