+++ src/ai_karen_engine/core/model_manager.py
