+++ src/ai_karen_engine/core/plugin_router.py
