+++ src/ai_karen_engine/core/prompt_router.py
@@ -1,3 +1,30 @@
-from src.core.prompt_router import PluginWrapper, PromptRouter
+import asyncio
+from ai_karen_engine.core.plugin_router import PluginRouter as BaseRouter
 
-__all__ = ["PluginWrapper", "PromptRouter"]
+
+class PluginWrapper:
+    def __init__(self, record):
+        self.manifest = record.manifest
+        self._handler = record.handler
+
+    def run(self, subtask: str):
+        """Execute the underlying plugin handler synchronously."""
+        return asyncio.run(self._handler({"subtask": subtask}))
+
+
+class PromptRouter:
+    """Very small router that always returns the autonomous_task_handler plugin."""
+
+    def __init__(self) -> None:
+        self.router = BaseRouter()
+
+    def route(self, _text: str) -> PluginWrapper:
+        record = self.router.get_plugin("autonomous_task_handler")
+        if not record:
+            raise ValueError("autonomous_task_handler plugin not found")
+        return PluginWrapper(record)
+
+    def generate_reply(self, text: str) -> str:
+        plugin = self.route(text)
+        result = plugin.run(text)
+        return str(result)
