+++ src/ai_karen_engine/core/reasoning/ice_integration.py
