+++ src/ai_karen_engine/core/soft_reasoning_engine.py
