--- src/core/__init__.py
@@ -1,6 +1,0 @@
-"""Core package initialization."""
-
-from .mesh_planner import MeshPlanner
-from .gpu_training import gpu_optimized_train
-
-__all__ = ["MeshPlanner", "gpu_optimized_train"]
