+++ tests/test_ice_integration.py
@@ -1,6 +1,6 @@
 import asyncio
 
-from src.core.reasoning.ice_integration import KariICEWrapper
+from ai_karen_engine.core.reasoning.ice_integration import KariICEWrapper
 
 
 def test_process_returns_keys():
