+++ tests/test_intent_engine.py
@@ -1,4 +1,4 @@
-from src.core.intent_engine import IntentEngine
+from ai_karen_engine.core.intent_engine import IntentEngine
 
 
 def test_detect_intent():
