+++ tests/test_llm_orchestrator.py
@@ -1,4 +1,4 @@
-from src.ai_karen_engine import SLMPool, LLMOrchestrator
+from ai_karen_engine import SLMPool, LLMOrchestrator
 from src.integrations.llm_utils import LLMUtils
