+++ tests/test_soft_reasoning.py
@@ -7,7 +7,7 @@ import asyncio
 
  
  
-from src.core.soft_reasoning_engine import SoftReasoningEngine
+from ai_karen_engine.core.soft_reasoning_engine import SoftReasoningEngine
 
 
 def test_ingest_and_query():
